+++ Assets/PROJECT/Scripts/GeneralHelpers/CoroutineWorkHorse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ace3db07e674b734481288b1cc3a6757
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
